IOS-6237: Get rid of duplicated `Decimal` extensions in the main app and `Visa` targets
