+++ Tangem/Modules/Send/Services/SendNotifications/SendNotificationManager.swift
@@ -8,6 +8,7 @@
 
 import Combine
 import BlockchainSdk
+import TangemFoundation
 
 protocol SendNotificationManagerInput {
     var feeValues: AnyPublisher<[SendFee], Never> { get }
